# OpenAPI.yaml
openapi: 3.0.0
info:
  title: store
  version: '1.0'
  description: A simple store
  contact:
    name: SecuritEase Dev
    url: 'https://www.securitease.com'
    email: internal@securitease.com
  termsOfService: 'https://www.securitease.com'
  license:
    name: Apache 2.0
    url: 'https://www.apache.org/licenses/LICENSE-2.0.htm'
servers:
  - url: 'http://localhost:8080'
    description: dev
paths:
  /order:
    get:
      summary: Get Orders
      tags:
        - store
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
      operationId: getOrders
      description: Get all orders
    post:
      summary: Create Order
      tags:
        - store
      responses:
        '201':
          description: Created
      operationId: createOrder
      description: Create a new order
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
  /order/{id}:
    get:
      summary: Get Order by ID
      tags:
        - Orders
      parameters:
        - name: id
          in: path
          required: true
          description: Order ID
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Order found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDTO'
        '404':
          description: Order not found
      operationId: getOrderById
      description: Retrieve a specific order by its ID
  /order/by-product/{productId}:
    get:
      summary: Get Orders by Product ID
      tags:
        - Orders
      parameters:
        - name: productId
          in: path
          required: true
          description: Product ID to search for
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Orders containing the specified product
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrderDTO'
      operationId: getOrdersByProductId
      description: Find all orders that contain a specific product
  /customer:
    get:
      summary: Get All Customers
      tags:
        - Customers
      responses:
        '200':
          description: Successfully retrieved all customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomerDTO'
      operationId: getAllCustomers
      description: Retrieve all customers
    post:
      summary: Create Customer
      tags:
        - Customers
      responses:
        '201':
          description: Customer created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerDTO'
        '400':
          description: Invalid input data
      operationId: createCustomer
      description: Create a new customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCustomerRequest'
  /customer/search:
    get:
      summary: Search Customers
      tags:
        - Customers
      parameters:
        - name: q
          in: query
          required: true
          description: Search query to match customer names (case-insensitive substring match)
          schema:
            type: string
            example: "john"
      responses:
        '200':
          description: Customers matching the search criteria
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomerDTO'
        '400':
          description: Missing or invalid search query
      operationId: searchCustomers
      description: Search for customers by name using case-insensitive substring matching
  /products:
    get:
      summary: Get All Products
      tags:
        - Products
      responses:
        '200':
          description: Successfully retrieved all products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductDTO'
      operationId: getAllProducts
      description: Retrieve all products
    post:
      summary: Create Product
      tags:
        - Products
      responses:
        '201':
          description: Product created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDTO'
        '400':
          description: Invalid input data
      operationId: createProduct
      description: Create a new product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProductRequest'
  /products/{id}:
    get:
      summary: Get Product by ID
      tags:
        - Products
      parameters:
        - name: id
          in: path
          required: true
          description: Product ID
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Product found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDTO'
        '404':
          description: Product not found
      operationId: getProductById
      description: Retrieve a specific product by its ID
components:
  schemas:
    OrderDTO:
      title: Order Response
      type: object
      required:
        - id
        - description
        - customer
        - productIds
      properties:
        id:
          type: integer
          format: int64
          description: Unique identifier for the order
          example: 1
        description:
          type: string
          description: Description of the order
          example: "Electronics Bundle Order"
        customer:
          $ref: '#/components/schemas/OrderCustomerDTO'
        productIds:
          type: array
          items:
            type: integer
            format: int64
          description: List of product IDs contained in this order
          example: [1, 2, 3]
          minItems: 1
    
    OrderCustomerDTO:
      title: Customer in Order
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: integer
          format: int64
          description: Customer ID
          example: 1
        name:
          type: string
          description: Customer name
          example: "John Doe"
    
    CustomerDTO:
      title: Customer Response
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: integer
          format: int64
          description: Unique identifier for the customer
          example: 1
        name:
          type: string
          description: Customer's full name
          example: "John Doe"
    
    ProductDTO:
      title: Product Response
      type: object
      required:
        - id
        - description
      properties:
        id:
          type: integer
          format: int64
          description: Unique identifier for the product
          example: 1
        description:
          type: string
          description: Product description
          example: "MacBook Pro 16-inch"
    
    CreateOrderRequest:
      title: Create Order Request
      type: object
      required:
        - description
        - customer
        - productIds
      properties:
        description:
          type: string
          description: Description of the order
          example: "Electronics Bundle Order"
        customer:
          type: object
          required:
            - id
          properties:
            id:
              type: integer
              format: int64
              description: ID of the customer placing the order
              example: 1
        productIds:
          type: array
          items:
            type: integer
            format: int64
          description: List of product IDs to include in the order
          example: [1, 2, 3]
          minItems: 1
    
    CreateCustomerRequest:
      title: Create Customer Request
      type: object
      required:
        - name
      properties:
        name:
          type: string
          description: Customer's full name
          example: "Jane Smith"
          minLength: 1
    
    CreateProductRequest:
      title: Create Product Request
      type: object
      required:
        - description
      properties:
        description:
          type: string
          description: Product description
          example: "iPhone 15 Pro"
          minLength: 1
    
    ErrorResponse:
      title: Error Response
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
          description: When the error occurred
        status:
          type: integer
          description: HTTP status code
          example: 404
        error:
          type: string
          description: Error type
          example: "Not Found"
        message:
          type: string
          description: Error message
          example: "Customer not found"
        path:
          type: string
          description: Request path that caused the error
          example: "/order"

tags:
  - name: Orders
    description: Order management operations
  - name: Customers  
    description: Customer management operations
  - name: Products
    description: Product management operations